+++ posthog/src/main/java/com/posthog/PostHog.kt
@@ -273,15 +273,20 @@ public class PostHog private constructor(
 
         val newDistinctId = distinctId ?: this.distinctId
 
+        val mergedProperties = buildProperties(
+            properties = properties,
+            userProperties = userProperties,
+            userPropertiesSetOnce = userPropertiesSetOnce,
+            groupProperties = groupProperties,
+        )
+
+        // sanitize the properties or fallback to the original properties
+        val sanitizedProperties = config?.propertiesSanitizer?.sanitize(mergedProperties.toMutableMap()) ?: mergedProperties
+
         val postHogEvent = PostHogEvent(
             event,
             newDistinctId,
-            properties = buildProperties(
-                properties = properties,
-                userProperties = userProperties,
-                userPropertiesSetOnce = userPropertiesSetOnce,
-                groupProperties = groupProperties,
-            ),
+            properties = sanitizedProperties,
         )
         queue?.add(postHogEvent)
     }
@@ -527,6 +532,13 @@ public class PostHog private constructor(
         return distinctId
     }
 
+    override fun debug(enable: Boolean) {
+        if (!isEnabled()) {
+            return
+        }
+        config?.debug = enable
+    }
+
     public companion object : PostHogInterface {
         private var shared: PostHogInterface = PostHog()
         private var defaultSharedInstance = shared
@@ -662,5 +674,8 @@ public class PostHog private constructor(
         }
 
         override fun distinctId(): String = shared.distinctId()
+        override fun debug(enable: Boolean) {
+            shared.debug(enable)
+        }
     }
 }